requirements.txt
Django>=5.2
psycopg2-binary
dj-database-url
google-genai